FF7: Refactor battle module into its folder and multiple files
Added a new namespace for battle module and split animations.cpp and animations.h into other files:
 - effect.cpp and effect.h that contains classes and method implementation related to effect decorators/wrappers
 - menu.cpp and menu.h that contains few functions related to battle menu
