+++ src/ff7/battle/camera.h
@@ -15,28 +15,31 @@
 
 #pragma once
 
-#include "../ff7.h"
+#include "../../common.h"
 
-class Camera
+namespace ff7::battle
 {
-    public:
-        Camera() = default;
-        ~Camera() = default;
+    class Camera
+    {
+        public:
+            Camera() = default;
+            ~Camera() = default;
 
-        void setRotationSpeed(float rotX, float rotY, float rotZ);
-        void setZoomSpeed(float speed);
-        void reset();
+            void setRotationSpeed(float rotX, float rotY, float rotZ);
+            void setZoomSpeed(float speed);
+            void reset();
 
-        void controlCamera(vector3<short>* cameraPosition, vector3<short>* cameraFocusPosition);
-    private:
-        vector3<float> rotationSpeed = { 0.0, 0.0, 0.0 };
-        vector3<float> rotationOffset = { 0.0, 0.0, 0.0 };
-        float zoomSpeed = 0.0f;
-        float zoomOffset = 0.0f;
-        const float minZoomDist = 5000.0f;
-        const float maxZoomDist = 30000.0f;
-        const float minVerticalAngle = 5.0f;
-        const float maxVerticalAngle = 85.0f;
-};
+            void controlCamera(vector3<short>* cameraPosition, vector3<short>* cameraFocusPosition);
+        private:
+            vector3<float> rotationSpeed = { 0.0, 0.0, 0.0 };
+            vector3<float> rotationOffset = { 0.0, 0.0, 0.0 };
+            float zoomSpeed = 0.0f;
+            float zoomOffset = 0.0f;
+            const float minZoomDist = 5000.0f;
+            const float maxZoomDist = 30000.0f;
+            const float minVerticalAngle = 5.0f;
+            const float maxVerticalAngle = 85.0f;
+    };
 
-extern Camera camera;
+    extern Camera camera;
+}
